+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [3.0.0](https://github.com/ErezNagar/lease-calculator/compare/v2.2.0...v3.0.0) - 08/7/2021
+
+### Fixed
+
+- Fixed getDiscountOffMsrpPercentage() to return null when Selling Price
+  - `getDriveOffPaymentDetails(): object[] | null`
+
 ## [2.2.0](https://github.com/ErezNagar/lease-calculator/compare/v2.1.0...v2.2.0) - 04/6/2021
 
 ### Added